chore: 인기 주류 관련 테스트 및 데이터 삭제
